COMMERCE-6005 Moved unit tests to the correct directory
